backend/requirements.txt
fastapi
uvicorn
boto3
passlib
email-validator
jose
python-dotenv